Progress on Photos and Tags

+++ src/main/java/szilveszterandras/vspf/handler/ValidateHandler.java
@@ -10,6 +10,7 @@ import com.google.gson.Gson;
 
 import szilveszterandras.vspf.dal.DAOFactory;
 import szilveszterandras.vspf.dal.Session;
+import szilveszterandras.vspf.payload.UserFilter;
 import szilveszterandras.vspf.payload.ValidateRequest;
 import szilveszterandras.vspf.payload.ValidateResponse;
 
@@ -24,24 +25,28 @@ public class ValidateHandler extends SimpleHandler<ValidateRequest> {
 	public ValidateHandler run() {
 		Session s = DAOFactory.getInstance().getSessionDAO().findByToken(this.payload.getToken());
 		Boolean isValid = s != null && (new Date()).compareTo(s.getExpiresAt()) <= 0;
+		UserFilter user = null;
 		if (isValid) {
 			// Update expiry of token
 			Date now = new Date();
 			Calendar expires = Calendar.getInstance();
 			expires.setTime(now);
-			expires.add(Calendar.SECOND, 10);
+			expires.add(Calendar.HOUR, 1);
 
 			s.setLastUpdated(now);
 			s.setExpiresAt(expires.getTime());
 
 			DAOFactory.getInstance().getSessionDAO().updateSession(s);
+			user = new UserFilter(DAOFactory.getInstance().getUserDAO().getUser(s.getUserId()));
 			this.connection.authenticate(s);
 		} else {
 			// Delete session
-			this.connection.dropSession();
-			DAOFactory.getInstance().getSessionDAO().deleteSession(s.getId());
+			if (s != null) {
+				this.connection.dropSession();
+				DAOFactory.getInstance().getSessionDAO().deleteSession(s.getId());
+			}
 		}
-		this.sendEvent((new Gson()).toJson(new ValidateResponse(isValid)));
+		this.sendEvent((new Gson()).toJson(new ValidateResponse(isValid, user)));
 		return this;
 	}
 }
